+++ Sources/StreamChatUI/CommonViews/InputTextView/InputTextView.swift
@@ -159,7 +159,8 @@ open class InputTextView: UITextView, ThemeProvider {
         addSubview(placeholderLabel)
         NSLayoutConstraint.activate([
             placeholderLabel.leadingAnchor.pin(equalTo: leadingAnchor, constant: directionalLayoutMargins.leading),
-            placeholderLabel.trailingAnchor.pin(lessThanOrEqualTo: trailingAnchor),
+            placeholderLabel.trailingAnchor.pin(equalTo: trailingAnchor, constant: -directionalLayoutMargins.trailing),
+            placeholderLabel.widthAnchor.pin(equalTo: layoutMarginsGuide.widthAnchor),
             placeholderLabel.topAnchor.pin(equalTo: topAnchor),
             placeholderLabel.bottomAnchor.pin(lessThanOrEqualTo: bottomAnchor),
             placeholderLabel.centerYAnchor.pin(equalTo: centerYAnchor)